requirements.txt
pycocotools
torch
torchvision
pillow
matplotlib
numpy
opencv-python
pytesseract
wand